Handle the no-ptrace.h case (I suppose that's only Solaris) gracefully.
We can't just drop KMonitorProcessIO.cpp from the sources, because
the class is referenced from the .ui files and from the .widget file.
That means that leaving it out causes linker errors with unresolved
symbols (like the constructor for KMonitorProcessIO).

So we need to produce a KMonitorProcessIO.cpp that is compatible but
doesn't use ptrace. This implementation just #ifdefs out everything
that is ptrace-dependent. That's ugly. A much nicer approach would
use a separate updater class that we hang in the UI, but that's
a larger reorganization of the code.

I tried producing a _noptrace version of the .ui and .widgets files,
but that promises to be a different kind of maintainence nightmare.

svn path=/trunk/KDE/kdebase/workspace/; revision=820310