Add a per-process CPU usage graph shown in the process list
Summary:
By adding a new PercentageHistoryRole returning a QVector with timestamps and
values, the delegate can show the history of the percentage as background.

This is currently only done for the CPU usage, but can easily be added for
memory as well.

The history vector is only created and filled after the first access on each
process.

Test Plan:
Added a small testcase which verifies that the history vector has an entry
with the current value after the first refresh after the initial access.
The end result looks like this:
https://phabricator.kde.org/file/data/mwuxhmkgb7p5gr2jl6ax/PHID-FILE-v76g3drarovozb4fue3w/KSysguard_with_per-process_cpu_graph

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9689