Port scripting to Qt WebEngine
Summary:
Unfortunately it relies on Qt WebKit features quite intensely, so a few
godawful hacks were necessary to make it work. Most of those are needed
to workaround restrictions of QWebChannel as bridge to the browser page.

We gain:
+ Independence from WebKit
+ Support for newer JS and HTML features in scripts

We lose:
- Performance (especially memory use)
- Independence from WebEngine
- Sanity

Test Plan:
Tested the detailed memory info pages on a few processes, info is
correct and the page seems to work fully.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D15209