Invert controls direction for process priority slider
Summary:
Since low 'nice' values correspond to high priority, the slider labels are inverted so that 'higher' corresponds to a lower nice number. However, this was not done for the slider control, resulting in the arrow keys moving the slider in the direction opposite to that expected by the user. This patch fixes this by setting invertedControls to 'true' for the priority slider.

BUG: 408074

Test Plan:
1. Open Ksysguard.
2. Go to the tab that lists processes.
3. Right click on a process and select 'Set Priority'.
4. Click on the slider that sets the CPU priority.
5. Press the left and right arrow keys on the keyboard and check if they move the slider in the appropriate direction.

Reviewers: cfeck

Reviewed By: cfeck

Subscribers: cfeck, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D21828