Create a plugin framework for processes
Summary:
Currently everything for processes is hardcoded with a method for each
process property. This is core functionality like CPU usage and memory
usage but it's not very extensible.

Currently ProcessModel is full of extra hacks to add X11 data when
really it should be a dumb proxier of information.

We have a pending patch to show network stats, and we have a pending
patch to add powertop information, which all work in a different way
from just reading data in /proc

In order to keep it flexible a more generic format method is added which
doesn't require hardcoding knowledge of types.

This patch is part of a series, next steps are adding various plugins,
stripping proces model - and then using the ProcessAttribute class to
provide the metadata for the core process attributes so that
ProcessModel can become a very simple view with no code duplication.

Reviewers: #plasma, ahiemstra

Subscribers: ahiemstra, alexde, broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23287