Move code to send signals and set priority of processes to a dedicated object
Summary:
This pulls the code out of KSysGuardProcessList and introduces it in a
new class called ProcessController. This should help make that code more
reusable and allow it to eventually be exposed to QML.

Test Plan:
Start ksysguard, try to send signals to processes and set priorities, including
root-owned processes. Everything should still work.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23566