Speed up sending request to ksysguardd
Summary:
While investigating bug 407526, I discovered that most of the time is
spent in SensorAgent::sendRequest.
This is because it searches the whole queue lineraly, which could take a
lot of time if there's a lot of requests. This patch speeds it up by
using a `QSet` of outstanding requests, obviating the need to search the
queues.

Part of {T11473}

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D23595