Introduce ProcessDataModel
Summary:
Last release a plugin system was introduced that allowed plugins to
provide columns of process data in a way that included enough metadata
to allow displaying of said data appropriately without the client
needing to be aware of the semantics of what that column represents.

This patch provides all process information in that new format. This is
then exposed as new, much simler, model.

This new model is designed to be consumable from a QML API for any
potential process data viewer.

Existing models are unchanged for maximum compatibility.

Test Plan: Used in another project

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: ahiemstra, broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D27509