Add a declarative plugin to ProcessCore
Summary:
This adds a QML plugin to ProcessCore that exposes ProcessController
and allows it to be used from QML, along with some changes the were
needed to actually use it from QML and a manual test to verify that
the QML bits work.

This will later on be expanded to also expose other types, including
ProcessDataModel from D27509

Test Plan: The test runs and can send signals to processes.

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28140