Use KSysGuard namespace for KSysGuard library targets
Summary:
Instead of using KF5 which is wrong since KSysGuard isn't part of
frameworks.

This needs a follow up patch to the ksysguard repo to make it use the new
namespace.

For backward compatibility, a KF5SysGuardConfig file is still installed that
aliases the new targets, but mentions they are deprecated.

Test Plan: Still builds

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29302