[sensors] Avoid breaking IDs of nodes within SensorTreeModel
item->name despite the name, is an ID of the relevant segment in the
node.

It is not the user facing name. That's always in d->sensorInfos.

By changing it here when double clicking an item we get a metaDataChange
as the Sensor object is constructed, which in turn breaks a lot of model
assumptions.

BUG: 421696