Add primaryActions and secondaryActions properties to SensorFace
This allows faces to expose a list of actions that can then be used by
an application to trigger behaviour that the face exposes. The primary
use case I have at the moment is exposing a set of actions that are
added to the toolbar in KSysGuardQML, but I can easily imagine more
uses.