Use shared instance of the process list
We might want to present the process list in several different views.
The process list is quite expensive, so if we ref-count the core we can
use it in multiple places.

Timers remain in their respecitive views as Processes has it's own
internal throttling mechanism.

Because that leaves us with a ref-counting issue on which plugin
attributes are enabled or not, that is changed to be implicit based on
whether we have dataChanged connected or not; which actually simplifies
some code.

Co-authored-by: David Redondo <kde@david-redondo.de>