Support converting "old" sensor ids to "new" sensor ids on face load
The network plugin changed the sensor ids of a number of sensors. To
support old configuration files, we need to translate these sensors from
their old IDs to the new. This adds support for that when loading sensor
faces and presets. In addition, when loading a pre-existing
configuration, it will write back the new sensor ids.