Handle smap read result in the correct thread
We set the context to "this" so we should auto disconnect when the
Processes object is destroyed.
But Qt::DirectConnection means that is run in the runner thread.

It is possible to have a scenario where "this" is active when the
connection is made, but destroyed before we get to this line.

BUG: 428160

But scoping the connect to Processes existing we would have leaked the
worker if Processes got destroyed whilst the worker was running.

CCBUG: 428048


(cherry picked from commit 9b5161cb9a16eaad205631b2a4c382dbbbe2072c)