Add API to Sensor and models to query current update rate
This allows us to change things to read the update rate from sensors
rather than needing to hardcode this. Right now the value for Sensor and
SensorDataModel is hardcoded to 500ms, which matches the new update rate
of KSystemStats. However, this new API allows us to change this without
needing to update everything else in the future.