Change Line Chart face X axis configuration to just a "history amount"
With the changes to have an API that determines what the update rate is
for sensors, it now becomes rather awkward to specify the number of
items a line chart should keep, since this number is dependant on the
update rate. So instead, change the line chart to have a single
configuration value indicating how much history to keep, in seconds.
This is then converted to a number of items based on the current update
rate of the sensor.