Write entries if face config changes
If we should not sync we correctly disable autosave. This disables writing to
the config and syncing it. However we still want to write to the config even
if the config should not be synced automatically, so when it is externally
synced the values are saved. This is the same method uses for the other config
properties, for example compare with setTitle().
BUG:429155
BUG:429167
BUG:429367