Do not insert non existing columns
d->sensors is what we derive columnCount() from, when the sensors are changed
or the model is reset we called beginInsertColumns for each sensor making the
outwards state of the model inconsistent. For example
when setting the sensors to a list of 4 sensors, beginInsertColumns would be
called 4 times, making some outside observers of the model think it would have
8 columns. So we need to initialize sensorInfos to avoid this.
Found when using the model with a QTransposeProxyModel.


(cherry picked from commit 74db33b1e0d166341850d162360f90d23f621c5f)