Allow using ExtendedLegend without setting a chart on it
This will be useful for SensorFaces that are not based on a QuickCharts chart.
We already implicitely require the type of sourceModel to be SensorDataModel
to format the sensor values. Make this explicit by stating the type. Use the data
provided by the model for the legend. The colors can be injected into the model
by the SensorFace to provide them to Legend. If "chart" is set, we continue
using it.