Do the sensor setters in two stages
When saveConfig() in ConfigSensors.qml is called we started resolving sensors,
when during this time loadConfig() was called we returned the old sensors, which
when then triggeded another saveConfig() call by  plasma-systemmonitor leading us to
effectively reset ourselves.
The setters are now in two stages, during resolving time we set the properties
to the unresolved value, so we are in sync with what the qml part did. Also we not
do another resolve when something tries to set the current unresolved ids.
BUG:432254