Only load plugins once
It makes no sense to instantiate each plugins once for each ExtendedProcesses
object. Instead a new ForwardingAttribute is introduced that simply mirrors
the information provided by the attributes of the plugins. This is needed to
correctly handle the enabled state of the attributes since it can be directly
manipulated from outside.