Set min/max for Y axis range spinboxes to minimum/maximum possible value
Some sensors have a range of gigabytes or terabytes, like memory and
disk. Since the base unit for these is bytes, we end up needing a very
large value in the range, which currently is not possible due to the
spinbox limiting the value. Internally, spinbox seems to use an int so
we can only store a value that fits within the range of an int, so we
use that as min/max value.

Longer term we want to have some ability to choose an actual "maximum"
sensor but that's going to need more work.

BUG: 433007
BUG: 424527