Properly fix column count and inserting new columns on metadata load
Commit 74db33b1 tried to fix this but did this through always creating
an instance of SensorInfo. This means we never end up inserting new data
in onMetaDataChanged, which breaks the "ready" state of the model and
can lead to an invalid value being produced for the "UpdateInterval"
model role, which means we end up with incorrect chart update rates.

We now use sensorInfos for determining column count and whether or not
the model is ready, which should better reflect the state of the model.

BUG: 433064


(cherry picked from commit 10d8d20a3bc4016ee271aee234f7727e5640cedc)