SensorDataModel: Do not insert columns that are out of range
It may happen that the metadata for a later column arrives before an
earlier one. In that case, we call beginInsertColumns with an out of
range column index.

This was verified to fix the the issue (and it still preserves the
correct order) by scrambling the metadata requests in `sensorsChanged()`
using std::shuffle and having the autotest still pass.

BUG: 433064