Assign a new color when configuring sensor colors
For some weird reason what we did had the effect of getting a reference
to the color of the ColorDialog. So when changing a sensor color it
would change the colors of all sensor colors that where configured
before.
BUG:434515
FIXED-IN:5.21.4


(cherry picked from commit ed89bbf919c0e2cf4b3a5a0c32978314e9e10e20)