Fix keyboard handling of preset selection
The view did not have focus correctly, due in part to this and a
kirigami issue.

We also want to prevent the plasma dialog from closing when escape is
pressed. Ideally we would have a magic solution in OverlaySheet, but
after we fixed focus it is impossible to do nicely.

shortcutOverride is only invoked on the item with real active focus, not
any parents.

BUG: 430563