Support specifying a unit for the Y range of Line charts
This makes use of the units from the current selected sensors to provide
a dropdown to select which unit to use for the range value. This is then
stored along with a multiplier in the config while and used in the line
chart to calculate the requested range for the chart.