Fix incorrect string termination in ConnectionMapping
After readlinkat, the nul character is not written at the correct index.
When readlinkat truncates the result, the code writes after the end of the array, which
triggers an assert in libstdc++.
In the other cases, it leaves an uninitialized char 'u' at the end of the array, but the
code stops at the ']' of "socket:[...]u".


(cherry picked from commit 6a60a626768ccafc51fbcca855bdea889fc4e5ab)