Handle process parent changes in ProcessDataModel
When the PPID of a process changes, it moves around in the model, changing the
layout. This needs to be announced properly, otherwise users of the model get
confused, leading to weird behaviour and crashes.

The added code is pretty much a direct copy from ProcessModel.

BUG: 446534


(cherry picked from commit a0d70929a1b5e38bd8bf61e1895321124acf03a7)