Port SensorFaceController to support JSON plugin metadata
Plasma has been ported to use JSON metadata everywhere. This means it is
not always guaranteed that a preset contains a "metadata.desktop" file.
So instead try to use KPackage::metadata() where possible. This also
changes savePreset to write out metadata in JSON format in addition to
the desktop file.