Fixed computational bug for bar chart spacing
The current calculation method for the bar chart spacing
would always result in a spacing size of zero (0).  This fix
implements a fixed size bar spacing scaling with the font size.
It allows the chart to scale well at any resolution, DPI and
font size as well as with any number of items charted.

BUG: 449868
FIXED-IN: 5.24.3