HorizontalBars: Use a smaller inner circle to meaningfully represent low values.
I noticed that the current version of the horizontal bars face does not meaningfully present any values that are lower than its own height, including zero, which is problematic for sensors that can be regularly expected to show low values, such as total CPU usage and swap space usage. For sensors like these, it would always draw a circle of width equal to its own height as a minimum value, which completely throws away any nuance at the low end, including the ability for the bar to represent zero at all. This problem is especially noticeable when you place a horizontal bars sensor on a thin vertical panel.