Faces: Do not write resolved sensors to preset file
The three sensor properties of FaceController contain the resolved
sensors. This means we are not writing group regular expressions to face
presets but are wiriting only the local sensors that were found for that
group, pretty much bypassing why we have the groups in the first place.
The unresolved sensors are only contained in the sensor configuration,
so read them from there and write that to the preset faceproperties.