Fix byte order handling in network helper
Network helper isn't consistent in using the host or network byte
order when dealing with IP addresses. This is a fix that changes all
data in memory to host byte order.

The problem that drew my attention to this was that while playing a
video on youtube, I could see no download rate in the system monitor
on the process. The reason was that youtube used ipv6 as its protocol
and ipv6 was particularly affected by the endianness problem. In my
view, it's clearer to use host endian wherever possible, even if
network byte order would usually do as well if applied consistently.