systemstats: Version DBus service/object/interface properly
In certain cases we may end up with mismatches if a newer client tries
to connect to an older ksystemstats or vice-versa. This may cause odd
issues like mismatching units. To avoid this, add a version number to
the appropriate DBus bits. This ensures we can change the version if
there are incompatible changes.

CCBUG: 477360