faces/horizontalbars: Use a fixed background rather than relying on style
The bars end up potentially being used in two separate styles, either
Plasma's styling or Desktop styling. This means we cannot rely on the
style for our background as it would be wrong in one of the two
contexts. So just hardcode a simple background that matches what we do
for Pie charts.

BUG: 480415