faces/pie: Allow center text to overlap the chart at small sizes
We now constrain the text so it never goes outside the central pie area,
but for small sizes this would cause things to wrap too early for
people's tastes. This change allows the text to overflow onto the pie
chart while still being constrained to the total applet size. When this
happens we render the text with an outline to improve contrast.

BUG: 488863

outline