faces/grid: Use CompactSensorFace for the compact rep
Going against my own comment, as it turns out that the size hints from
FaceGrid are not as useful as I hoped and are not really fixable either.
So instead just follow the general sizing hints we set up for compact
reps, with a slight tweak to the minimum size to account for the grid.

BUG: 489307