facepackages/colorgrid: use FormLayout as top level
Most configuration pages for faces use FormLayout as their top element,
but the one for colorgrid wraps it in a ColumnLayout. As a consequence,
this is the only page where the form is vertically centered. While this
does look nice, especially for a page with few options, the config pages
should be consistent with each other – if vertically centering is
desired, it should happen for all of them.

This change removes the ColumnLayout, making FormLayout the top level
item and removing the vertical centering.