faces/colorgrid: Replace grid implementation with the one from FaceGrid
This face was copied from an earlier version of FaceGrid. FaceGrid now
has an improved implementation that works a lot better in panels etc.
Copy that implementation over to ColorGrid so we get the same
improvements there.

Ultimately we should get rid of ColorGrid in favour of a simple colored
rectangle face that can be placed in FaceGrid, but that's for later.