processcore: Make ApplicationDataModel aggregate similar CGroups into one entry
This changes ApplicationDataModel to aggregate similar CGroups into one
application entry.

The primary motivation for this is that a single Flatpak application
instance may spawn several separate CGroups that we would list as
separate applications before this change, but which logically belong to
the same application. Upstream has indicated this behaviour is
intentional and seems to have no inclination on adding anything to allow
us to identify these cases, so we are left with just merging everything.

This is also slightly more aggressive in filtering out some things that
are currently listed but aren't really applications, most notably some
D-Bus autostarted applications.

BUG: 481551