faces: Ensure the temporary dir for a preset remains until installed
QTemporaryDir deletes the temporary directory when it goes out of scope.
However, installing the preset with KPackage is an asynchronous
operation that can extend beyond the current scope. This can cause the
preset data to be deleted while we're still installing. To ensure this
doesn't happen, extend the lifetime of the QTemporaryDir into the lambda
we use to handle the job finished, so that it lives at least as long as
the job.

BUG: 485164