processcore: Convert smaps reading to use a single long-lived thread
Instead of creating a separate runnable that we run on a thread pool.

This gives us more control over how often we read smaps_rollup, allowing
us to tweak the amount of CPU usage from reading that file. The thread
will sleep for a specified duration each cycle, then read smaps_rollup
for a number of processes.

This reduces the rate at which memory information is updated, but since
that does not change as quickly as CPU usage, this seems like a
reasonable tradeoff.