processcore: Use MemoryInfo for tracking memory information in Process
This reworks how memory information is collected and tracked, with the
intent to both reduce CPU usage and increase precision of memory
information.

On Linux, we have two sources of memory information, /proc/PID/stat(m)
and /proc/PID/smaps_rollup. The former is cheap to read but imprecise,
the latter is expensive to retrieve but more accurate. This is now
reflected in the MemoryInfo struct that keeps track of both imprecise
and precise information as well as the delta between them. The delta is
used as a correction factor on the imprecise data, which is assumed to
be updated at a higher rate than the precise data.

This also reduces the rate at which precise data is collected on Linux,
leading to a reduction in CPU usage after the initial data is retrieved.