faces/Choices: fix resetting tree model on close
When the sensor filer/browser popup closes, the model is reset by
setting it to the parent index. This doesn't actually work properly:
when the model is on the third level, it will only go back to the second
level instead of the top, and if the model is already at the (invalid)
top index, an error message will be printed as an invalid index has no
parent.

Instead, only go to the parent of a valid node, and continue upward
until it's no longer valid.