faces/ExtendedLegend: handle null source model
If configured to not show a legend, piechart and other faces pass a null
sourceModel to ExtendedLegend. ExtendedLegend reads the sensors property
of the sourceModel as part of its default properties. which causes an
unneccessary type error if it's null.

Instead, use optional chaining to avoid the error.


(cherry picked from commit f42fcdd23b70fd2414b297709492bc93f0143395)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>