Drop Process::macContext()
The implementation does a silly thing. It reads the data into a 1000 byte buffer
and converts that to string, even though most of that buffer will be null bytes.

This means we allocate a lot of memory (~1% of all allocations of plasma-systemmonitor
startup).

Then the data isn't even used at all, so just drop the entire thing