faces: set maximum width for arbitrary input spinboxes
The angle spinboxes for piechart and the duration spinbox for linechart
allow arbitrary input, but use default (implicit) sizing. This means
that user input is not constrained by the from/to settings of the
spinbox, and the width of the control can grow without bounds, even
beyond the constraints of the window. Such large inputs are generally
pointless (and will likely fail anyway when the spinbox input is
parsed). But we can fix the graphical glitch to look less broken.

This change sets a maximum value for the length of these spinboxes that
matches the fixed-width spinboxes used on their pages.

BUG: 467813
FIXED-IN: 6.4.0