processcore: Rewrite ProcessesLinux::getIOStatistics to be less confusing
Apparently this was reading the entire file and then assuming a certain
order of fields. This was pretty hard to follow code that broke easily
because it was using a different pattern from the rest of the functions.
Change it to read line by line and check which field we are reading.
This should make things a fair bit more robust and easier to read.

BUG: 504435