processcore: Rewrite ProcessesLinux::getIOStatistics to be less confusing
Apparently this was reading the entire file and then assuming a certain
order of fields. This was pretty hard to follow code that broke easily
because it was using a different pattern from the rest of the functions.
Change it to read line by line and check which field we are reading.
This should make things a fair bit more robust and easier to read.

BUG: 504435


(cherry picked from commit 8ed933e643598b85c67203c586a2fec5fc9ee8d5)

Co-authored-by: Arjen Hiemstra <ahiemstra@heimr.nl>