faces/horizontalbars: set default color if colorSource not loaded yet
When the Disk Usage plasma widget is added to a panel or desktop, the
qml is initialized before everything on the backend is loaded. This
causes a warning when trying to assign the color bars, which are still
undefined.

Instead, set them to some color (white chosen at semi-random) as a
default; when the backend is fully loaded they'll get replaced by the
proper values.


(cherry picked from commit 8c5cf79f8f474c025625c5e73543f20bc6188fc2)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>