processcore: Batch changes to processes in ProcessDataModel
We are currently emitting a single data change signal for every change
in a process, even if it is a change to different columns of the same
row. This is rather inefficient and can cause multiple updates of a view
without that being necessary.

Instead, rather than directly emitting dataChanged when a process
changes, record that the process changed and process all changes after
processes have been updated. This allows us to go through all the
changes and figure out what the range is of changed rows and columns and
emit a single dataChanged for that entire range.

This hopefully also fixes a crash in TreeView that can rarely happen
when ProcessDataModel emits dataChanged

CCBUG: 502826