Achieve WCAG AA for axis labels
Part of https://invent.kde.org/plasma/plasma-desktop/-/issues/149

The contrast ratio here was 2.42:1, which doesn't meet WCAG AA, as it
requires 4.5:1 or higher. This is caused by using the disabled text
color, which is not semantically correct here since these labels aren't
actually disabled (only disabled items are allowed to have lower
contrast according to WCAG).

Instead, use 0.75 opacity for the default text color, which is what
we've standardized on elsewhere. This raises the contrast ratio to:

- 5.73:1 on the opaque gray window background color
- 5.44:1 on a translucent background in Plasma with the current default
wallpaper (though this will vary a bit with different backgrounds)

Now they meet WCAG AA.