processcore: Explicitly stop capturing packets on network helper terminate
In some cases, we may end up in a situation where we're terminating with
low or no network traffic. In such a case, we would block indefinitely
since there was no mechanism to notify `Capture::nextPacket()` that we
want to terminate.

This changes things so that we explicitly call `stop()` when the main
thread quits, which closes the pcap source and signals the wait
condition in Capture that we should no longer wait.

BUG: 507883