SensorFaceController: Possibility to reload the config from a given group
When a systemmonitor plasmoid is dragged between the desktop and panel,
all its configuration gets moved to a new group which lives under the
panel containment config group.
Therefore the SensorFaceController config doesn't live on the same
group anymore, and when calling reloadConfig, the actual groups used
need to change to the new ones.

CCBUG:508184